Scale-test harness + f32 edge storage: 2x memory headroom at 25k
Two interlocking pieces:

1. Seeded-fixture test harness (new crates/seed-corpus/)

   Deterministic adversarial-corpus builders: star(n), chain(n), ba(n,m).
   Zero external deps (xorshift RNG). Emits markdown so WikiBackend::open
   exercises the full scan -> tfidf -> graph -> persist path.

   crates/wiki-backend/tests/runtime_audit.rs gains two parameterised tests:
     - fixtures: invariant matrix across 7 built-in shapes. Row-stochastic
       residual, no self-loops, positive costs, budget respect, sigmoid
       contraction, convergence-flag honesty, >=1/2 OPT via exact DP.
     - fixtures_scale: perf-envelope ladder ba-{1k,2.5k,5k} by default,
       plus ba-{10k,25k} under AUDIT_SCALE=heavy. Reports build_ms,
       query p50/p99, RSS delta, iteration count.

   Supports AUDIT_FIXTURE=star-100 | chain-500 | ba-1000-4 to drive a
   single shape for bisecting regressions.

2. f32 edge storage + WikiIndex raw_weights deduplication

   ScoredGraph stores raw, adj, and adj_t_values as f32 instead of f64.
   SpMV accumulates in f64 to keep rounding well below the Lipschitz bound.
   is_row_stochastic tolerance relaxed 1e-9 -> 1e-5 to match the ~7-digit
   f32 mantissa. All dependent crates (spread, temporal-graph, gauntlet)
   updated to cast at write/read boundaries; spread's inner loop also
   hoists the `next` and `thresholded` scratch buffers out of the
   iteration loop (was 2 allocs per iter, now zero).

   WikiIndex drops the `raw_weights: Vec<f64>` field that mirrored
   ScoredGraph's storage. persist::save now takes &ScoredGraph and reads
   weights live; persist::load returns (WikiIndex, ScoredGraph). On-disk
   blob format stays f64 for backward compat with existing engine.db files.

Measured on the ba-25000-8 rung of the heavy ladder:
  RSS:        11.2 GB -> 5.8 GB   (-48%)
  build_ms:    27,406 -> 15,867   (-42%)
  query p99:   62 ms  -> 60 ms    (noise)
  iters/conv:  unchanged (math preserved)

All 136 workspace tests pass, plus the 7 fixture invariants and 5 scale
rungs. The runtime-audit harness is the load-bearing proof that the math
still holds under halved-precision storage.

Next structural levers (not in this change): drop dense raw+adj matrices
entirely (keep CSR only), mmap-backed CSR, inverted-index TF-IDF,
incremental rebuild.

+++ crates/gauntlet/src/chaos.rs
@@ -237,7 +237,7 @@ pub fn vopr_run(seed: u64, n: usize, n_steps: usize) -> Result<(), String> {
                 let dst = rng.next_usize(n);
                 if src != dst {
                     let raw = graph.raw_matrix_mut();
-                    raw[src * n + dst] = rng.next_f64() * 2.0;
+                    raw[src * n + dst] = (rng.next_f64() * 2.0) as f32;
                     graph.renormalize();
                 }
                 check_invariants(&graph, &label)?;
@@ -265,7 +265,7 @@ pub fn vopr_run(seed: u64, n: usize, n_steps: usize) -> Result<(), String> {
                         _ => rng.next_f64() * 100.0,   // moderately large
                     };
                     let raw = graph.raw_matrix_mut();
-                    raw[src * n + dst] = extreme;
+                    raw[src * n + dst] = extreme as f32;
                     graph.renormalize();
                 }
                 check_invariants(&graph, &label)?;
